rename patient slice to patientData
