+++ App.js
@@ -3,12 +3,12 @@ import { Provider } from 'react-redux';
 import { combineReducers, configureStore } from '@reduxjs/toolkit';
 
 import RootNavigator from './src/navigation/RootNavigator';
-import patientReducer from './src/features/patient/patientSlice';
+import patientDataReducer from './src/features/patient/patientDataSlice';
 import authReducer from './src/features/auth/authSlice';
 
 const rootReducer = combineReducers({
   auth: authReducer,
-  patient: patientReducer,
+  patient: patientDataReducer,
 });
 
 const store = configureStore({
